requirements.txt
asttokens==2.0.8
backcall==0.2.0
backports.functools-lru-cache==1.6.4
bcrypt==4.0.1
brotlipy==0.7.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
comtypes==1.1.13
cryptography==38.0.3
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==0.10.0
fonttools==4.34.4
idna==3.3
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter_core==4.11.1
kiwisolver==1.4.4
matplotlib==3.5.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
numpy==1.23.2
opencv-python==4.6.0.66
packaging==21.3
paramiko==2.11.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pip==22.3
playsound==1.3.0
prompt-toolkit==3.0.30
psutil==5.9.1
pure-eval==0.2.2
pycaw==20220416
pycparser==2.21
Pygments==2.13.0
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
pywin32==303
pyzmq==23.2.1
requests==2.28.1
setuptools==65.2.0
six==1.16.0
stack-data==0.4.0
torch==1.12.1
torchaudio==0.12.1
torchsummary==1.5.1
torchvision==0.13.1
tornado==6.2
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.11
wcwidth==0.2.5
wheel==0.37.1
win-inet-pton==1.1.0